Project reorg.
