+++ src/main/scala-2.11/markets/actors/participants/strategies/CancellationStrategy.scala
@@ -1,0 +1,12 @@
+package markets.actors.participants.strategies
+
+import markets.orders.Order
+
+import scala.collection.mutable
+
+
+trait CancellationStrategy {
+
+  def cancelOneOf[T <: mutable.Iterable[Order]](outstandingOrders: T): Option[Order]
+
+}
