--- src/main/scala-2.11/markets/participants/strategies/CancellationStrategy.scala
@@ -1,12 +1,0 @@
-package markets.participants.strategies
-
-import markets.orders.Order
-
-import scala.collection.mutable
-
-
-trait CancellationStrategy {
-
-  def cancelOneOf[T <: mutable.Iterable[Order]](outstandingOrders: T): Option[Order]
-
-}
