--- src/test/scala-2.11/markets/participants/OrderIssuerSpec.scala
@@ -1,102 +1,0 @@
-/*
-Copyright 2016 David R. Pugh
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-   http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-package markets.participants
-
-import akka.actor.{ActorRef, ActorSystem}
-import akka.agent.Agent
-import akka.testkit.{TestKit, TestProbe}
-
-import markets.MarketsTestKit
-import markets.orders.limit.{LimitAskOrder, LimitBidOrder}
-import markets.orders.market.{MarketAskOrder, MarketBidOrder}
-import markets.participants.strategies.TestTradingStrategy
-import markets.tickers.Tick
-import markets.tradables.{TestTradable, Tradable}
-import org.scalatest.{FeatureSpecLike, GivenWhenThen, Matchers}
-
-
-class OrderIssuerSpec extends TestKit(ActorSystem("OrderIssuerSpec"))
-  with MarketsTestKit
-  with FeatureSpecLike
-  with GivenWhenThen
-  with Matchers {
-
-  /** Shutdown TestSystem after running tests. */
-  def afterAll(): Unit = {
-    system.terminate()
-  }
-
-  val tradable = TestTradable("GOOG")
-
-  val market = TestProbe()
-  val markets = Map[Tradable, ActorRef](tradable -> market.ref)
-
-  val initialTick = Tick(1, 1, 1, 1, timestamp())
-  val tickers = Map[Tradable, Agent[Tick]](tradable -> Agent(initialTick)(system.dispatcher))
-
-  feature("An OrderIssuer should be able to issue limit orders.") {
-
-    val tradingStrategy = new TestTradingStrategy(Some(1), 1)
-    val props = TestOrderIssuer.props(markets, tickers, tradingStrategy)
-    val orderIssuer = system.actorOf(props)
-
-    scenario("A OrderIssuer receives a SubmitAskOrder message.") {
-
-      When("an OrderIssuer receives a SubmitAskOrder message...")
-      orderIssuer.tell(SubmitAskOrder, testActor)
-
-      Then("...the market should receive an ask order.")
-      market.expectMsgAnyClassOf(classOf[LimitAskOrder])
-
-    }
-
-    scenario("A OrderIssuer receives a SubmitBidOrder message.") {
-
-      When("an OrderIssuer receives a SubmitBidOrder message...")
-      orderIssuer.tell(SubmitBidOrder, testActor)
-
-      Then("...the market should receive an bid order.")
-      market.expectMsgAnyClassOf(classOf[LimitBidOrder])
-    }
-  }
-
-  feature("An OrderIssuer should be able to issue market orders.") {
-
-    val tradingStrategy = new TestTradingStrategy(None, 1)
-    val props = TestOrderIssuer.props(markets, tickers, tradingStrategy)
-    val orderIssuer = system.actorOf(props)
-
-    scenario("A OrderIssuer receives a SubmitAskOrder message.") {
-
-      When("an OrderIssuer receives a SubmitAskOrder message...")
-      orderIssuer.tell(SubmitAskOrder, testActor)
-
-      Then("...the market should receive an ask order.")
-      market.expectMsgAnyClassOf(classOf[MarketAskOrder])
-
-    }
-
-    scenario("A OrderIssuer receives a SubmitBidOrder message.") {
-
-      When("an OrderIssuer receives a SubmitBidOrder message...")
-      orderIssuer.tell(SubmitBidOrder, testActor)
-
-      Then("...the market should receive an bid order.")
-      market.expectMsgAnyClassOf(classOf[MarketBidOrder])
-    }
-  }
-
-}
